Clean streeteasy address by removing extra spaces
I don't know what else words().join(' ') was intended to do. In
fact, I'm only assuming it was meant to do this. What I know it
was doing is replacing '-'s with spaces, so 26-22 4th Street
got turned to 26 22 4th Street. 2622 4th Street is the same
as the address we want, at least in Astoria, but 26 22 4th Street
is different and does not exist.